Extract attachPolicies to own file in src/aws
