+++ src/aws/attachPolicies.js
@@ -1,0 +1,23 @@
+const retryAsync = require("../util/retryAsync");
+const { iam } = require('./services');
+
+const attachPolicies = (policyArns, roleName) => {
+  const attachPolicyPromises = policyArns.map((policyArn) => {
+    const policyParams = {
+      PolicyArn: policyArn,
+      RoleName: roleName,
+    };
+
+    return retryAsync(
+      () => {
+        return iam.attachRolePolicy(policyParams).promise();
+      },
+      2,
+      1000
+    );
+  });
+
+  return Promise.all(attachPolicyPromises);
+};
+
+module.exports = attachPolicies;
